root: refactored nativeCI for a standardized library
Like it or not, PowerShell is superbly painful to deal with and
Windows is the only one that does not play nicely compared to Apple
and other UNIX systems. I best refactor the nativeCI library for
a standardized one for documentable and maintainable sakes. Hence,
let's do this.

This patch refactors nativeCI for a standardized library in root
directory.

Signed-off-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>