srcPYTHON: updated Start CI job recipe to issue manual instruction
Since the Python venv requires manual actions to configure the
terminal and AutomataCI are executable and not sourcable, we
should update the start CI job instruction to issue the manual
instruction instead. Hence, let's do this.

This patch updates Start CI job recipe to issue manual instruction
in srcPYTHON/ directory.

Co-authored-by: Galyna, Cory <124406765+corygalyna@users.noreply.github.com>
Co-authored-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>
Signed-off-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>