automataCI: performed major code clean up for stability
At the moment, the automataCI libraries' codes are too dirty and
as we're closing to Release job recipes development, it's better to
perform the first major code clean up for stable future development.
Hence, let's do this.

This patch performs major code clean up for stability for future
development in automataCI/ directory.

Co-authored-by: Galyna, Cory <124406765+corygalyna@users.noreply.github.com>
Co-authored-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>
Signed-off-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>