automataCI: added GPG package crypto-signing for release job
To ensure there is a redundancy across the release packages'
integrity, it's best we also include in the crypto-signing using
GPG capability prior to shasum checking. This way, the customers
can have options to verify a distributed product is safe and from
the right origin.

This patch adds GPG package crypto-signing capability for release
job in automataCI/ directory.

Co-authored-by: Galyna, Cory <124406765+corygalyna@users.noreply.github.com>
Co-authored-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>
Signed-off-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>