automataCI: fixed critical Release Job bug nuking the repo
Appearently, the GIT::hard_reset_to_init function behaves in a way
that if the releases/ directory is not a git repository, it shall
nuke the root repo back to init commit and destroys everything
locally and remotely. This is a critical security vulenrability
requiring immediate attention. Hence, let's do this.

This patch fixes critical Release Job bug nuking the entire repo
in automataCI/ directory.

Discovered-by: Galyna, Cory <124406765+corygalyna@users.noreply.github.com>
Co-authored-by: Galyna, Cory <124406765+corygalyna@users.noreply.github.com>
Co-authored-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>
Signed-off-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>