automataCI: added simulation mechanism to test the Release Job API
To make sure the most important and most risky release job, we best
add simulation run capability into it and permits mock run for
testing purposes. That way, it shall not be left out and do crazy
matters anymore.

This patch adds simulation mechanism to test the Release Job API
in automataCI/ directory.

Co-authored-by: Galyna, Cory <124406765+corygalyna@users.noreply.github.com>
Co-authored-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>
Signed-off-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>